// constants/generatedData.ts
export const productImages: Record<string, ImageSourcePropType> = {
  'absolute-zero': require('../assets/products/absolute-zero.png'),
  'babyliss-chameleon': require('../assets/products/absolute-zero.png'),
  'babyliss-super-collection': require('../assets/products/absolute-zero.png'),
  'babyliss-super-skeleton': require('../assets/products/absolute-zero.png'),
  'blade-ice': require('../assets/products/blade-ice.png'),
  'boosted': require('../assets/products/boosted.png'),
  'cape': require('../assets/products/cape.png'),
  'ergo': require('../assets/products/ergo.png'),
  'fade-brush': require('../assets/products/fade-brush.png'),
  'ff2020t': require('../assets/products/ff2020t.png'),
  'fresh-fade-2020c': require('../assets/products/fresh-fade-2020c.png'),
  'gamma-boosted': require('../assets/products/absolute-zero.png'),
  'gamma-boosted-trimmer': require('../assets/products/absolute-zero.png'),
  'gamma-boosted-up': require('../assets/products/absolute-zero.png'),
  'gold-fx': require('../assets/products/gold-fx.png'),
  'gtx-z': require('../assets/products/gtx-z.png'),
  'instinct': require('../assets/products/absolute-zero.png'),
  'instinct-metal': require('../assets/products/absolute-zero.png'),
  'jrl-ghost': require('../assets/products/absolute-zero.png'),
  'jrl-ghost-kit': require('../assets/products/absolute-zero.png'),
  'jrl-lamborghini-black': require('../assets/products/absolute-zero.png'),
  'jrl-onyx': require('../assets/products/absolute-zero.png'),
  'magic-clip': require('../assets/products/magic-clip.png'),
  'magic-clip-black': require('../assets/products/absolute-zero.png'),
  'magic-clip-gold': require('../assets/products/absolute-zero.png'),
  'omni-anomaly': require('../assets/products/absolute-zero.png'),
  'omni-arcane': require('../assets/products/absolute-zero.png'),
  'power-clip': require('../assets/products/power-clip.png'),
  'prodigy': require('../assets/products/prodigy.png'),
  'saber': require('../assets/products/saber.png'),
  'saber-2-black': require('../assets/products/absolute-zero.png'),
  'saber-trimmer-black': require('../assets/products/absolute-zero.png'),
  'saber-trimmer-white': require('../assets/products/absolute-zero.png'),
  'skeleton-fx': require('../assets/products/skeleton-fx.png'),
  'stinger-guards': require('../assets/products/stinger-guards.png'),
  'stylecraft-rebel': require('../assets/products/absolute-zero.png'),
  'wahl-hi-viz': require('../assets/products/absolute-zero.png'),
  'wahl-senior': require('../assets/products/wahl-senior.png'),
  'wahl-senior-2': require('../assets/products/absolute-zero.png'),
  'wahl-vapor': require('../assets/products/absolute-zero.png'),
  'x-ergo': require('../assets/products/x-ergo.png'),
};

export const brandLogos: Record<string, ImageSourcePropType> = {
  'babyliss': require('../assets/brands/babyliss.png'),
  'wahl': require('../assets/brands/wahl.png'),
  'andis': require('../assets/brands/andis.png'),
  'gamma-plus': require('../assets/brands/gamma-plus.png'),
  'stylecraft': require('../assets/brands/stylecraft.png'),
  'jrl': require('../assets/brands/jrl.png'),
  'tomb45': require('../assets/brands/tomb45.png'),
  'supreme-trimmer': require('../assets/brands/supreme-trimmer.png'),
  'omni': require('../assets/brands/andis.png'),
};

export const generatedBrands: Brand[] = [
  {
    "id": "1",
    "name": "BaByliss PRO",
    "logo": brandLogos['babyliss']
  },
  {
    "id": "2",
    "name": "Wahl",
    "logo": brandLogos['wahl']
  },
  {
    "id": "3",
    "name": "Andis",
    "logo": brandLogos['andis']
  },
  {
    "id": "4",
    "name": "Gamma+",
    "logo": brandLogos['gamma-plus']
  },
  {
    "id": "5",
    "name": "StyleCraft",
    "logo": brandLogos['stylecraft']
  },
  {
    "id": "6",
    "name": "JRL",
    "logo": brandLogos['jrl']
  },
  {
    "id": "7",
    "name": "Tomb45",
    "logo": brandLogos['tomb45']
  },
  {
    "id": "8",
    "name": "Supreme Trimmer",
    "logo": brandLogos['supreme-trimmer']
  },
  {
    "id": "9",
    "name": "Omni",
    "logo": brandLogos['omni']
  }
];

export const generatedProducts: Product[] = [
  {
    "id": "1",
    "name": "Skeleton FX",
    "brand": "BaByliss PRO",
    "category": "Trimmers",
    "price": 149.99,
    "image": productImages['skeleton-fx'],
    "stores": [
      {
        "name": "Salons Direct",
        "url": "https://www.salonsdirect.com/products/babyliss-pro-snap-fx-skeleton-dlc-blade",
        "inStock": true,
        "price": 144.99
      },
      {
        "name": "Coolblades",
        "url": "https://www.coolblades.co.uk/babyliss-pro-super-motor-skeleton-trimmer/",
        "inStock": true,
        "price": 149.99
      },
      {
        "name": "Chris & Sons",
        "url": "https://chrisandsons.co.uk/skeleton-fx",
        "inStock": true,
        "price": 142.5
      },
      {
        "name": "Amazon UK",
        "url": "https://amazon.co.uk/dp/skeleton-fx",
        "inStock": true,
        "price": 154
      }
    ],
    "trending": true,
    "isNew": false,
    "description": "Lightweight skeleton-style outliner trimmer with a precision T-blade. A barber shop staple.",
    "colors": [
      {
        "name": "Black",
        "hex": "#1A1A1A",
        "image": productImages['skeleton-fx']
      },
      {
        "name": "Rose Gold",
        "hex": "#E0BFB8",
        "image": productImages['gold-fx']
      },
      {
        "name": "Chrome",
        "hex": "#C0C0C0",
        "image": productImages['skeleton-fx']
      }
    ]
  },
// ...
];